+++ README.md
@@ -1,16 +1,8 @@
- 
-
 ![ynodesk](https://user-images.githubusercontent.com/2998216/201456135-270da105-a4fa-4976-a69a-3a69e5d3fe59.png)
-
-  
 
 A desktop client for [Yume Nikki Online](https://ynoproject.net/) with Discord Rich Presence. Show your friends what game you're playing and what room you're in. Also looks pretty cool as an installed program.
 
-  
-
 Created as a simple Electron wrapper around the website, with `discord-rpc` and some hacks to make save exporting/importing work.
-
-  
 
 ## Features
 
@@ -20,15 +12,11 @@ Created as a simple Electron wrapper around the website, with `discord-rpc` and 
 
 - Smooth integration with Yume Nikki Online.
 
-  
-
 ## Rich Presence
 
 ![Rich Presence Example 1](https://user-images.githubusercontent.com/2998216/201456282-6337d763-db5c-4fc2-b399-00b3513b1f7b.png)
 
 ![Rich Presence Example 2](https://user-images.githubusercontent.com/2998216/201456297-8cb36ebb-6400-4ae8-9804-ce51bcf3c1b5.png)
-
-  
 
 ## Download
 
@@ -36,81 +24,63 @@ The releases are currently Windows-only.
 
 [Click here to go to the download page for the latest release.](https://github.com/joaovitorbf/ynodesktop/releases/latest)
 
-  
-
 ## Development Setup
 
 To set up the project locally, you need to have [Node.js](https://nodejs.org/) and [Yarn](https://yarnpkg.com/) installed. Then follow these steps:
 
-  
-
 1. Clone the repository:
 
 ```sh
-
 git clone https://github.com/joaovitorbf/ynodesktop.git
 
 cd ynodesktop
-
 ```
-
-  
 
 2. Install dependencies:
 
 ```sh
-
 yarn install
-
 ```
-
-  
 
 3. Run the application:
 
 ```sh
-
 yarn start
-
 ```
-
-  
 
 ## Building the Project
 
 To build the project, use the following command:
 
 ```sh
-
 yarn dist
-
 ```
 
 This will create the distributable files in the `dist` directory.
 
-  
-
 ## Project Structure
 
--  `main.js`: The main entry point of the application.
+- `main.js`: The main entry point of the application.
 
--  `package.json`: Contains project metadata and dependencies.
+- `preload.js`: A script that runs before the main process, often used to set up initial configurations or preload certain resources.
 
--  `yarn.lock`: Dependency lock file to ensure consistent setups.
+- `scripts/discordRpcUtils.js`: Utility functions for handling Discord RPC (Remote Procedure Call) integrations.
 
-  
+- `scripts/promptinjection.js`: Handles prompt injection logic for the application.
+
+- `scripts/titlebar.js`: Manages the title bar functionality and customization.
+
+- `package.json`: Contains project metadata and dependencies.
+
+- `yarn.lock`: Dependency lock file to ensure consistent setups.
 
 ## Contributing
 
 Feel free to fork the repository and submit pull requests. For major changes, please open an issue first to discuss what you would like to change.
 
-  
-
 ## License
 
 This project is licensed under the MIT License. See the [LICENSE](LICENSE) file for details.
-
-  
 
 ## YNOproject
