+++ package.json
@@ -2,7 +2,7 @@
   "name": "ynodesktop",
   "version": "1.1.9",
   "description": "A Yume Nikki Online desktop client with Discord Rich Presence",
-  "main": "main.js",
+  "main": "src/main.js",
   "scripts": {
     "start": "electron .",
     "pack": "electron-builder --dir",
