+++ src/main.js
@@ -1,9 +1,17 @@
-const { app, BrowserWindow, session, ipcMain, dialog, shell } = require("electron");
+const {
+  app,
+  BrowserWindow,
+  session,
+  ipcMain,
+  dialog,
+  shell,
+} = require("electron");
 const contextMenu = require("electron-context-menu");
 const DiscordRPC = require("discord-rpc");
 const Store = require("electron-store");
-const promptInjection = require("./promptinjection");
-const titlebar = require("./titlebar");
+const promptInjection = require("./scripts/promptinjection");
+const titlebar = require("./scripts/titlebar");
+const { updatePresence } = require("./scripts/discordRpcUtils");
 const path = require("path");
 
 const store = new Store();
@@ -79,34 +87,6 @@ contextMenu({
 var client = new DiscordRPC.Client({ transport: "ipc" });
 client.login({ clientId: "1028080411772977212" }).catch(console.error);
 
-const mappedIcons = [
-  "amillusion",
-  "answeredprayers",
-  "braingirl",
-  "deepdreams",
-  "flow",
-  "someday",
-  "unevendream",
-  "yno-logo",
-  "yume2kki",
-  "yumenikki",
-  "mumarope",
-  "genie",
-  "mikan",
-  "2kki",
-  "yume",
-  "prayers",
-  "muma",
-  "dreamgenie",
-  "mikanmuzou",
-  "ultraviolet",
-  "sheawaits",
-  "oversomnia",
-  "yumetsushin",
-  "nostalgic",
-  "collectiveunconscious",
-];
-
 const createWindow = () => {
   const win = new BrowserWindow({
     width: 1052,
@@ -118,7 +98,7 @@ const createWindow = () => {
     titleBarStyle: "hidden",
     webPreferences: {
       preload: path.join(__dirname, "preload.js"),
-      backgroundThrottling: false
+      backgroundThrottling: false,
     },
   });
   var loopInterval = null;
@@ -141,8 +121,8 @@ const createWindow = () => {
       document.querySelector('#content')?.scrollTo(0,0)}`); // Disable scroll ingame
   });
 
-  win.webContents.on('devtools-opened', () => {
-    win.webContents.send('log-app-version', app.getVersion());
+  win.webContents.on("devtools-opened", () => {
+    win.webContents.send("log-app-version", app.getVersion());
   });
 
   // open wiki links in a web browser for better readability, but open maps in a new electron window
@@ -200,148 +180,14 @@ function clientLoop(win) {
     let splitTitle = title.split(" Online ");
     if (splitTitle[1]?.trim() == "- YNOproject") {
       if (splitTitle[0].trim() == "ゆめ2っき") {
-        updatePresence(web, "Yume 2kki");
+        updatePresence(web, client, "Yume 2kki");
       } else {
-        updatePresence(web, splitTitle[0].trim());
+        updatePresence(web, client, splitTitle[0].trim());
       }
     } else {
-      updatePresence(web);
+      updatePresence(web, client);
     }
   });
-}
-
-function isConnected(text) {
-  let privatemodes = [
-    "وضع الخاص",
-    "Private Mode",
-    "Private Mode",
-    "Private Mode",
-    "Mode privé",
-    "Private Mode",
-    "プライベートモード",
-    "비공 개 모드",
-    "Tryb prywatny",
-    "Modo Privado",
-    "Private Mode",
-    "Приватный режим",
-    "Private Mode",
-    "Riêng tư",
-    "私密模式",
-  ];
-  let connecteds = [
-    "متصل",
-    "Verbunden",
-    "Connected",
-    "Conectado",
-    "Connecté(e)",
-    "Connesso",
-    "接続済み",
-    "연결됨",
-    "Połączony",
-    "Conectado",
-    "Conectat",
-    "В сети",
-    "Bağlı",
-    "Đã kết nối",
-    "已连接",
-  ];
-
-  if (privatemodes.includes(text) || connecteds.includes(text)) {
-    return true;
-  }
-  return false;
-}
-
-function retryConnection(err) {
-  console.log("Retry IPC");
-  console.log(err);
-  client = new DiscordRPC.Client({ transport: "ipc" });
-  client.login({ clientId: "1028080411772977212" }).catch(console.error);
-}
-
-function updatePresence(web, gamename = null) {
-  web.executeJavaScript("window.onbeforeunload=null;");
-  console.log("Update Presence");
-  console.log(gamename);
-
-  if (gamename == null) {
-    client
-      .setActivity({
-        largeImageKey: "yno-logo",
-        largeImageText: "Yume Nikki Online Project",
-        details: "Choosing a game...",
-        instance: false,
-        buttons: [{ label: "Play YNOproject", url: `https://ynoproject.net/` }],
-      })
-      .catch(retryConnection);
-  } else {
-    web
-      .executeJavaScript(
-        `
-        (function() {
-          var querys = document.querySelector("#locationText").querySelectorAll("span") // This seems to be the new location format for C.U.
-          if (querys == null || querys.length < 1) {
-            querys = document.querySelector("#locationText").querySelectorAll("a") // and this is the old format for other games
-          }
-
-          if (querys == null || querys.length < 1)
-          {
-            currentLocation = "Unknown";
-          }
-          else
-          {
-            for (i = 0; i < querys.length; i++)
-            {
-              isPowerOfTwo = (i % 2) == 0;
-              currentLocation = (!isPowerOfTwo ? "(" : "") + querys.item(i).textContent + (!isPowerOfTwo ? ")" + (i + 1 == querys.length ? "" : " |") : "") + " ";
-            }
-          }
-
-          return {
-            name: window.location.pathname.replaceAll('/', ''),
-            currentLocation,
-            connected: document.querySelector("#connStatusText")?.textContent,
-            url: document.URL
-          }
-        })()
-      `
-      )
-      .then((data) => {
-        const condensedName = gamename
-          .toLowerCase()
-          .replace(" ", "")
-          .replace(".", "");
-        let buttonGame = gamename;
-
-        // There is a length limit for activity buttons of 32 chars
-        // so we need to shorten the game name
-        if (gamename == "Collective Unconscious") {
-          buttonGame = "C.U.";
-        } else if (gamename.length > 32) {
-          buttonGame = gamename.slice(0, 28) + "...";
-        }
-
-        let activityButtons = [
-          { label: "Play " + buttonGame + " online", url: data.url },
-        ];
-        client
-          .setActivity({
-            largeImageKey: mappedIcons.includes(condensedName)
-              ? condensedName + "-icon"
-              : `https://ynoproject.net/images/door_${data.name}.gif`,
-            largeImageText: gamename,
-            smallImageKey: "yno-logo",
-            smallImageText: "Yume Nikki Online Project",
-            details: "Dreaming on " + gamename,
-            state: isConnected(data.connected)
-              ? data.currentLocation
-              : "Disconnected",
-            instance: false,
-            buttons: activityButtons,
-          })
-          .catch(retryConnection);
-      });
-  }
 }
 
 function saveSession() {
