+++ src/scripts/titlebar.js
